feat(components) Created a separate file for the Content component to avoid repetition
