+++ pages/index.js
@@ -27,6 +27,7 @@ import StepCard from '@/components/Molecules/StepCard';
 import StarIcon from '@/components/Atoms/StarIcon';
 import ClientsIcon from '@/components/Atoms/ClientsIcon';
 import DolarIcon from '@/components/Atoms/DolarIcon';
+import Content from '@/components/Molecules/Content';
 
 const iconsMapping = {
   star: <StarIcon />,
@@ -45,23 +46,6 @@ function getVideoTransformationsWithReactVideo() {
 
   return myVideo;
 }
-
-const Content = ({ content }) => {
-  const { title, subtitle, description } = content;
-  return (
-    <div className="px-4 text-left">
-      <h2 className="mb-6 text-2xl font-bold text-purple display-font">
-        {title}
-      </h2>
-      <h3 className="lg:text-[38px] leading-tight text-3xl font-bold display-font text-purple mb-4">
-        {subtitle}
-      </h3>
-      <p className="text-2xl font-semibold lg:text-2xl text-dark-grey">
-        {description}
-      </p>
-    </div>
-  );
-};
 
 const Home = ({ data }) => {
   const router = useRouter();
